requirements/requirements-dev.txt
pytest-django==4.7.0
behavex==3.2.0
django-debug-toolbar==4.2.0
drf-yasg==1.21.7
pytest-cov==4.1.0
pytest-html==4.1.1
pytest-mock==3.12.0
pytest-sugar==1.0.0
pytest-instafail==0.5.0
pytest-xdist==3.5.0